Small changes to remove some of these functionalities from the backend to the app
